requirements.txt
torch
torchvision
Pillow
flask
flask_cors